pyproject.toml
[tool.poetry]
name = "pycrawler"
version = "0.1.0"
description = ""
authors = ["codemker <ciderifz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.9.4"
beautifulsoup4 = "^4.12.2"
selenium = "^4.16.0"
pillow = "^10.1.0"
crawler = "^0.0.2"
pytest = "^7.4.4"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
